\name{ecpc-package}
\alias{ecpc-package}
\alias{ecpc-package}
\docType{package}
\title{
\packageTitle{ecpc}
}
\description{
\packageDescription{ecpc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecpc}
\packageIndices{ecpc}

See \code{\link{ecpc}} for example code.
}
\author{
\packageAuthor{ecpc}

Maintainer: \packageMaintainer{ecpc}
}

