\name{ecpc-package}
\alias{ecpc-package}
\alias{ecpc-package}
\docType{package}
\title{
\packageTitle{ecpc}
}
\description{
\packageDescription{ecpc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ecpc}
\packageIndices{ecpc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ecpc}

Maintainer: \packageMaintainer{ecpc}
}
\references{
}
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
